backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pymupdf==1.23.8
sentence-transformers==2.3.1
chromadb==0.4.22
groq==0.4.2
python-dotenv==1.0.0
langchain-text-splitters==0.0.1
numpy<2.0
transformers==4.36.2
# Install torch CPU version separately:
# pip install torch --index-url https://download.pytorch.org/whl/cpu
torch==2.1.2
